+++ util-core/src/test/scala/com/twitter/io/BufReaderTest.scala
@@ -28,7 +28,9 @@ class BufReaderTest extends FunSuite with Checkers {
       n < 0 ||
       bytes.length == 0 ||
       Await
-        .ready(r.read(n), 5.seconds).poll.exists(_.throwable.isInstanceOf[Reader.ReaderDiscarded])
+        .ready(r.read(n), 5.seconds).poll.exists(
+          _.throwable.isInstanceOf[ReaderDiscardedException]
+        )
     }
   }
 }