+++ CHANGELOG.rst
@@ -7,10 +7,16 @@ Note that ``PHAB_ID=#`` and ``RB_ID=#`` correspond to associated messages in com
 Unreleased
 ----------
 
+Breaking API Changes
+~~~~~~~~~~~~~~~~~~~~
+
 * util-core: `c.t.u.Future.raiseWithin` methods now take the timeout exception as a call-by-name
   parameter instead of a strict exception. While Scala programs should compile as usual, Java
   users will need to use a `scala.Function0` as the second parameter. The helper
   `c.t.u.Function.func0` can be helpful. ``PHAB_ID=D229559``
+
+* util-core: Rename `c.t.io.Reader.ReaderDiscarded` to `c.t.io.ReaderDiscardedException`.
+  ``PHAB_ID=D231969``
 
 18.10.0
 -------